refactor: io table cell
