+++ web/src/components/trace/IOPreview.tsx
@@ -1,4 +1,4 @@
-import { JSONView } from "@/src/components/ui/code";
+import { JSONView } from "@/src/components/ui/CodeJsonViewer";
 import { z } from "zod";
 import { deepParseJson } from "@/src/utils/json";
 import { cn } from "@/src/utils/tailwind";
