+++ web/src/features/public-api/components/ApiKeyList.tsx
@@ -1,7 +1,7 @@
 import Header from "@/src/components/layouts/header";
 import { Button } from "@/src/components/ui/button";
 import { Card } from "@/src/components/ui/card";
-import { CodeView } from "@/src/components/ui/code";
+import { CodeView } from "@/src/components/ui/CodeJsonViewer";
 import {
   Dialog,
   DialogContent,
